review: move detect_gridded into lumen.ai.utils
Per review on #1823: detect_gridded is a small one-function helper;
it does not warrant its own module. Moved into lumen/ai/utils.py
alongside the other AI helpers and removed the standalone gridded.py.

Updated imports in lumen/ai/agents/base_view.py and
lumen/tests/ai/test_gridded.py.
